+++ chemfusekit/lr.py
@@ -7,11 +7,14 @@ import pandas as pd
 from sklearn.linear_model import LogisticRegression
 
 from chemfusekit.__utils import run_split_test, print_confusion_matrix, print_table, GraphMode
+from .__base import BaseSettings, BaseDataModel, BaseClassifier
+from .pca import PCADataModel
 
-class LRSettings:
+
+class LRSettings(BaseSettings):
     '''Holds the settings for the LR object.'''
-    def __init__(self, algorithm: str = 'liblinear', output: GraphMode = GraphMode.NONE,
-                 test_split: bool = False):
+    def __init__(self, algorithm: str = 'liblinear', output: GraphMode = GraphMode.NONE, test_split: bool = False):
+        super().__init__(output, test_split)
         if algorithm not in [
             'lbfgs',
             'liblinear',
@@ -21,22 +24,17 @@ class LRSettings:
             'saga'
         ]:
             raise ValueError(f"{algorithm}: this algorithm does not exist.")
-        if test_split is True and output is GraphMode.NONE:
-            raise Warning(
-                "You selected test_split but it won't run because you disabled the output."
-            )
         self.algorithm = algorithm
-        self.output = output
-        self.test_split = test_split
 
 
-class LR:
+class LR(BaseClassifier):
     '''Class to store the data, methods and artifacts for Logistic Regression'''
-    def __init__(self, settings: LRSettings, array_scores: np.ndarray, y: np.ndarray):
-        self.settings = settings
-        self.array_scores = array_scores
-        self.y = y
-        self.model: Optional[LogisticRegression] = None
+    def __init__(self, settings: LRSettings, data: BaseDataModel):
+        super().__init__(settings, data)
+        if isinstance(data, PCADataModel):
+            self.array_scores = data.array_scores
+        else:
+            self.array_scores = data.x_train.drop('Substance', axis=1).values
 
     def lr(self):
         '''Performs Logistic Regression'''
@@ -47,10 +45,10 @@ class LR:
             random_state=0,
             class_weight='balanced',
             max_iter=10000
-        ).fit(self.array_scores, self.y)
+        ).fit(self.array_scores, self.data.y)
 
         # See the classes the model used
-        classes = np.unique(self.y)
+        classes = np.unique(self.data.y)
         classes = classes.reshape((1, len(classes)))
         coefficients = model.coef_.transpose()
         intercepts = model.intercept_.reshape((1, len(model.intercept_)))
@@ -72,12 +70,12 @@ class LR:
         predictions = model.predict(self.array_scores)
 
         # This tells us the accuracy of our model in calibration
-        scores = model.score(self.array_scores, self.y)
+        scores = model.score(self.array_scores, self.data.y)
 
         # Save the trained model
         self.model = model
 
-        sample_column = self.y.reshape((1, self.y.shape[0]))
+        sample_column = self.data.y.reshape((1, self.data.y.shape[0]))
         pred_column = predictions.reshape((1, predictions.shape[0]))
         prob_column = probabilities.transpose()
 
@@ -93,7 +91,7 @@ class LR:
         )
 
         print_confusion_matrix(
-            self.y,
+            self.data.y,
             predictions,
             "Confusion Matrix based on whole data set",
             self.settings.output
@@ -109,7 +107,7 @@ class LR:
 
             run_split_test(
                 self.array_scores,
-                self.y,
+                self.data.y,
                 split_model,
                 extended=True,
                 mode=self.settings.output
@@ -117,10 +115,7 @@ class LR:
 
     def predict(self, x_sample: pd.DataFrame):
         '''Performs LR prediction once the model is trained.'''
-        if self.model is None:
-            raise RuntimeError("The LR model is not trained yet!")
-
-        prediction = self.model.predict(x_sample)
+        prediction = super().predict(x_sample)
         probabilities = self.model.predict_proba(x_sample)
 
         classes = self.model.classes_.reshape((self.model.classes_.shape[0], ))