+++ tests/test_knn.py
@@ -5,7 +5,7 @@ import pandas as pd
 import numpy as np
 
 from chemfusekit.knn import KNNSettings, KNN, GraphMode
-from chemfusekit.lldf import LLDFSettings, LLDF, LLDFModel, Table
+from chemfusekit.lldf import LLDFSettings, LLDF, LLDFDataModel, Table
 
 
 class TestKNN(unittest.TestCase):
@@ -78,11 +78,11 @@ class TestKNN(unittest.TestCase):
             sheet_name="Sheet1",
             preprocessing="none"
         )
-        lldf = LLDF([table1, table2], lldf_settings)
+        lldf = LLDF(lldf_settings, [table1, table2])
         lldf.lldf()
 
         # settings parameter
-        wrong_settings = LLDFModel(pd.DataFrame([1]), pd.DataFrame([1]), np.asarray([1]))
+        wrong_settings = LLDFDataModel(pd.DataFrame([1]), pd.DataFrame([1]), np.asarray([1]))
         with self.assertRaises(TypeError):
             KNN(wrong_settings, lldf.fused_data)  # pass an object of the wrong class as settings
 
@@ -106,7 +106,7 @@ class TestKNN(unittest.TestCase):
             sheet_name="Sheet1",
             preprocessing="none"
         )
-        lldf = LLDF([table1, table2], lldf_settings)
+        lldf = LLDF(lldf_settings, [table1, table2])
         lldf.lldf()
 
         # Set up and run KNN (no output)
@@ -138,7 +138,7 @@ class TestKNN(unittest.TestCase):
             sheet_name="Sheet1",
             preprocessing="none"
         )
-        lldf = LLDF([table1, table2], lldf_settings)
+        lldf = LLDF(lldf_settings, [table1, table2])
         lldf.lldf()
 
         # Set up KNN without training it